requirements.txt
aiohttp==3.9.4rc0; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
attrs==23.2.0; python_version >= '3.7'
automat==22.10.0
bitarray==2.9.2
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
constantly==23.10.4; python_version >= '3.8'
cytoolz==0.12.3; implementation_name == 'cpython'
eth-abi==4.2.1; python_version < '4' and python_full_version >= '3.7.2'
eth-account==0.8.0; python_version >= '3.6' and python_version < '4'
eth-hash[pycryptodome]==0.6.0; python_version >= '3.8' and python_version < '4'
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0; python_version >= '3.7' and python_version < '4'
eth-typing==3.5.2; python_version < '4' and python_full_version >= '3.7.2'
eth-utils==2.3.1; python_version >= '3.7' and python_version < '4'
frozenlist==1.4.1; python_version >= '3.8'
hexbytes==0.3.1; python_version >= '3.7' and python_version < '4'
hyperlink==21.0.0
idna==3.6; python_version >= '3.5'
incremental==22.10.0
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
lru-dict==1.2.0
multidict==6.0.5; python_version >= '3.7'
parsimonious==0.9.0
protobuf==5.26.0rc2; python_version >= '3.8'
pycryptodome==3.20.0
python-statemachine==2.1.2; python_version < '3.13' and python_version >= '3.7'
pyunormalize==15.1.0; python_version >= '3.6'
referencing==0.33.0; python_version >= '3.8'
regex==2023.12.25; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
rlp==3.0.0
rpds-py==0.18.0; python_version >= '3.8'
setuptools==69.1.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.12.1; python_version >= '3.7'
twisted==24.2.0rc1; python_full_version >= '3.8.0'
typing-extensions==4.10.0rc1; python_version >= '3.8'
urllib3==1.26.18; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
web3==6.15.1; python_full_version >= '3.7.2'
websockets==12.0; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
zope-interface==6.2; python_version >= '3.7'
